Refatorar estrutura modular da interface
